Xcos: fix AddressSanitizer reported issues
The FROMWSB block had an out-of-memory access (one index) when using the `"NOT_A_KNOT method"`. This has been fixed, and a unit test for all parameters was added.

When compiling, c_pass2 used ctree3 compiled function which had an out-of-memory access (one index). The C-code implementation was used to check Scilab behaviour on `unique()` and `gsort()`; the C-code was removed.

See https://gitlab.com/scilab/scilab/-/merge_requests/1550