force close subprocess launched by host() when Scilab is closed or killed
Launch Scilab, then execute:
```
[s,o,e]=host(fullfile(SCI,"bin","scilab") + " -nw -e ""sleep(60000)"" -quit")
```

Killing the main Scilab must also destroy the one launched using `host()`.

## Windows

When creating a child process with `CreateProcessW()`, the child will normally continue running even if the parent process is killed.  
To ensure it terminates when the parent does, a **job object** must be created and the child process assigned to it.  
When the parent process exits or closes the job handle, all child processes in that job are automatically terminated.

## Linux & macOS

When using `fork()` and `exec()`, a child process normally continues running after the parent process dies.  
`fork()` creates a new process in the parent group, `execl` replaces that process with the command, in this case `sh`.  
`sh` will interpret the user command and create a new process in the same process group.  
So, watching `SIGTERM` in Scilab to forward it to its process group will kill all processes created using `host()`.

#### The Scilab script (bin/scilab)
Running Scilab via the `scilab` script starts two processes: one for the script itself and one for the Scilab binary.  
Killing the script alone will not kill the binary. By using `exec` in the script, the Scilab binary **replaces the script**, so only one process remains.

See https://gitlab.com/scilab/scilab/-/merge_requests/1581