test_run return the exit code when launched in a pipe
## Actual status
- Many Scilab tests use `then, pause` when the test must fail.
- `test_run` replaces them with `bugmes();quit;` before the test execution.
- So, all `.dia.ref` of tests with `then, pause` are generated with `bugmes();quit;`
- some tests contains hard-coded `bugmes();quit;`
- All tests that contain `then, pause` must be stopped on pause when executed with `exec`

## Changes
- Fix hard-coded `bugmes();quit;` in tests.
- Fix some `then, pause` that are not replaced by `test_run`.
- Add `exit(...)` in `bugmes` and `assert_generror` to ensure that the `host` command executed in `test_run` will exit with a positive status.
- Keep the replacement by `bugmes();quit;` to avoid changing all `.dia.ref`

## %onprompt
`%onprompt` is used to check `lasterror()`, write a message in the diary, then exit with status 0.
`%onprompt` is called at the Scilab prompt display, but when Scilab is launched with `-quit` and **not** in a pipe (test_run on a local machine), the console thread is not started, so no prompt implies that `%onprompt` is never called.
Because of this, the exit status of tests can differ because Scilab starts in a pipe on the CI but not on the local machine.
Now, Scilab always checks `lasterror()` at exit to return the error status; `%onprompt` is no longer needed.

## Checks this MR
Local machine test running:
```
./bin/scilab-cli -e "test_run development_tools 14743" -quit
./bin/scilab-cli -e "test_run development_tools test_run_error" -quit
```

CI like test running:
```
echo | ./bin/scilab-cli -e "test_run development_tools 14743" -quit
echo | ./bin/scilab-cli -e "test_run development_tools test_run_error" -quit
```

See https://gitlab.com/scilab/scilab/-/merge_requests/1554